Move TCB into kernel core
