+++ kernel/src/threading/scheduler.rs
@@ -10,7 +10,8 @@ use core::mem::{ManuallyDrop, MaybeUninit};
 use core::num::NonZero;
 use core::ops::{Deref, DerefMut};
 use core::ptr::NonNull;
-use crate::hal::{self, ThreadControlBlock, ThreadState};
+use crate::hal;
+use crate::threading::tcb::{ThreadControlBlock, ThreadState};
 use core::sync::atomic::{AtomicUsize, Ordering};
 #[cfg(feature = "preemptive")] use core::time::Duration;
 use kernel_api::memory::physical::highmem;
